requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Vector embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# LLM integration
langchain-google-genai==0.0.4
langchain-core==0.1.4
langchain-openai==0.0.3

# API and web interface
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pydantic==2.4.2

# MongoDB
pymongo==4.6.0

# Utilities
tqdm==4.66.1
matplotlib==3.8.1
seaborn==0.13.0 